Straightened Router impl.
